Make config.js, format code and remove redundant props
