+++ src/components/Contacts/SelectedContact.module.css
@@ -37,36 +37,37 @@
   border-bottom: none;
 }
 
-.contact-info, .main-info {
+.contact-info,
+.main-info {
   display: flex;
   flex-direction: column;
 }
 
-.contact-info{
+.contact-info {
   background: rgb(56, 56, 56);
   padding: 1.5rem 1rem;
   border-radius: 10px;
   width: 95%;
 }
 
-.main-info{
+.main-info {
   text-align: center;
-  gap: .5rem;
+  gap: 0.5rem;
 }
 
-.actions{
+.actions {
   display: flex;
   gap: 3rem;
 }
 
-.actions button{
+.actions button {
   cursor: pointer;
   background: none;
   border: none;
   color: rgb(77, 76, 76);
 }
 
-.actions button:hover{
+.actions button:hover {
   color: #ccc;
   transform: scale(1.1);
 }
